requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow
pillow==10.0.0
numpy>=1.26.0
gunicorn==21.2.0
python-dotenv==1.0.1
